(filament_swirl) Updated to remove mapping context from domain
