+++ applications/clawpack/advection/2d/filament_swirl/swirl/swirl_user.h
@@ -31,9 +31,6 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #ifdef __cplusplus
 extern "C"
 {
-#if 0
-}
-#endif
 #endif
 
 typedef struct swirl_options
@@ -55,12 +52,13 @@ void swirl_options_store (fclaw2d_global_t* glob, swirl_options_t* user);
 
 const swirl_options_t* swirl_get_options(fclaw2d_global_t* glob);
 
-fclaw2d_domain_t* swirl_create_domain(sc_MPI_Comm mpicomm, fclaw_options_t* gparms);
+/* Swirl */
+void swirl_create_domain(fclaw2d_global_t *glob);
+void swirl_initialize(fclaw2d_global_t* glob);
+void swirl_finalize(fclaw2d_global_t* glob);
+
 
 #ifdef __cplusplus
-#if 0
-{
-#endif
 }
 #endif
